services/process-video-large/requirements.txt
Flask==3.*
gunicorn==21.*
google-cloud-storage==2.*
google-cloud-firestore==2.*
